- refactor: move vis-overview-panel into separate component - style: create separate stylesheet for overview panel
