+++ src/css/VisOverviewPanel.css
@@ -1,0 +1,171 @@
+.vis-overview-panel {
+    position: absolute;
+    width: 100%;
+    height: 100%;
+    border-right: 1px solid lightgray;
+    background-color: white;
+    overflow-y: none;
+    z-index: 999;
+    left: 0px;
+    transition-property: left, display;
+    transition-duration: 0.2s;
+    transition-timing-function: ease;
+    transition-behavior: allow-discrete;
+
+    .navigation-container {
+        display: flex;
+        align-items: center;
+        padding: 0px 20px;
+        font-size: 18px;
+        border-bottom: 1px solid lightgray;
+        background: #f6f6f6;
+
+        .config-button {
+            svg {
+                vertical-align: unset;
+            }
+        }
+        .config-button:hover {
+            background-color: #eaeaea;
+        }
+        .config-button:active {
+            background-color: #e3e3e3;
+        }
+
+        .title-github-link:hover,
+        .title-doc-link:hover,
+        .feedback a:hover {
+            background-color: #eaeaea;
+        }
+
+        .title-github-link:active,
+        .title-doc-link:active,
+        .feedback a:active {
+            background-color: #e3e3e3;
+        }
+    }
+
+    .title {
+        padding: 10px;
+        font-size: 18px;
+        border-bottom: 1px solid lightgray;
+        background: #f6f6f6;
+    }
+
+    .button {
+        width: 30px;
+        height: 30px;
+        cursor: pointer;
+        position: absolute;
+        right: 5px;
+        top: 12px;
+    }
+
+    .title small {
+        font-weight: normal;
+    }
+
+    .overview-root {
+        display: flex;
+        flex-direction: column;
+        margin-top: 50px;
+        gap: 0px;
+
+        * {
+            box-sizing: border-box !important;
+        }
+
+        .overview-container {
+            display: flex;
+            flex-wrap: wrap;
+            flex: 1;
+            width: 100%;
+            padding: 20px;
+            gap: 20px;
+            overflow-y: auto;
+        }
+
+        .overview-header {
+            display: flex;
+            flex-direction: row;
+            width: 100%;
+            height: 100px;
+            border-right: 1px solid lightgrey;
+            overflow: hidden;
+            background-color: #f3f9ff;
+            border-bottom: 1px solid #aabacd;
+
+            .menu-container {
+                width: auto;
+            }
+
+            .cancer-selector {
+            }
+        }
+
+        .overview-status {
+            background: rgba(210, 210, 210, 0.1);
+            height: 25px;
+            float: left;
+            border-bottom: 1px solid lightgrey;
+            padding: 0px;
+            margin: 0px;
+            color: grey;
+            align-content: middle;
+            display: flex;
+            justify-content: center;
+        }
+    }
+
+    table {
+        width: 100%;
+        border-spacing: 0px 0px;
+        border-collapse: collapse;
+    }
+
+    tr /* :nth-child(even) */ {
+        border-top: 1px solid lightgray;
+    }
+
+    td {
+        border-radius: 0px;
+        padding-left: 90px !important;
+    }
+
+    .overview-title {
+        position: absolute;
+        left: -10px;
+        top: 120px;
+        transform: rotate(-90deg);
+        font-weight: bold;
+        font-size: 18px;
+        text-shadow: 0px 0px 12px #f6f6f6;
+    }
+
+    .overview-container > div {
+        padding: 4px;
+    }
+
+    .selected-overview,
+    .unselected-overview {
+        display: block;
+        text-align: center;
+        flex: 1;
+    }
+
+    .selected-overview {
+        box-shadow: inset 0px 0px 0px 2px #2e94ff;
+    }
+
+    .unselected-overview {
+        cursor: pointer;
+    }
+
+    .unselected-overview:hover {
+        box-shadow: inset 0px 0px 0px 2px lightgray;
+    }
+
+    td {
+        padding: 20px;
+    }
+}
